before rest api

+++ controllers/admin.js
@@ -1,18 +1,50 @@
 const Product = require('../models/product');
+const { validationResult } = require('express-validator/check');
+const fileHelper = require('../util/file');
 
 exports.getAddProduct = (req, res, next) => {
     res.render('admin/edit-product', {
         docTitle: 'Add Product',
         path: '/admin/add-product',
         editing: false,
+        hasError: false,
+        errorMessage: null,
+        validationErrors: []
     });
 };
 
 exports.postAddProduct = (req, res, next) => {
     const title = req.body.title; // req.body.["name" field dans l'input du form]
-    const imageUrl = req.body.imageUrl;
+    const image = req.file;
     const price = req.body.price;
     const description = req.body.description;
+
+    if (!image) {
+        return res.status(422).render('admin/edit-product', {
+            docTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: { title: title, price: price, description: description },
+            errorMessage: 'Attached file is of the wrong type.',
+            validationErrors: []
+        });
+    }
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).render('admin/edit-product', {
+            docTitle: 'Add Product',
+            path: '/admin/add-product',
+            editing: false,
+            hasError: true,
+            product: { title: title, price: price, description: description },
+            errorMessage: errors.array()[0].msg,
+            validationErrors: errors.array()
+        });
+    }
+
+    const imageUrl = image.path;
+
     const product = new Product({
         title: title,
         price: price,
@@ -24,7 +56,11 @@ exports.postAddProduct = (req, res, next) => {
         .then(() => {
             res.redirect('/admin/products');
         })
-        .catch();
+        .catch((err) => {
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error); // next avec une erreur en argument va skip tous les middleware et aller directement a un middleware de gestion d'erreur (dans app.js: app.use((error, req, res, next) => {}))
+        });
 };
 
 exports.getEditProduct = (req, res, next) => {
@@ -43,6 +79,9 @@ exports.getEditProduct = (req, res, next) => {
                 path: '/admin/edit-product',
                 editing: editMode,
                 product: product,
+                hasError: false,
+                errorMessage: null,
+                validationErrors: []
             });
         });
 };
@@ -51,8 +90,21 @@ exports.postEditProduct = (req, res, next) => {
     const prodId = req.body.productId;
     const updatedTitle = req.body.title;
     const updatedPrice = req.body.price;
-    const updatedImageUrl = req.body.imageUrl;
+    const image = req.file;
     const updatedDesc = req.body.description;
+
+    const errors = validationResult(req);
+    if (!errors.isEmpty()) {
+        return res.status(422).render('admin/edit-product', {
+            docTitle: 'Edit Product',
+            path: '/admin/edit-product',
+            editing: true,
+            hasError: true,
+            product: { title: updatedTitle, price: updatedPrice, description: updatedDesc, _id: prodId },
+            errorMessage: errors.array()[0].msg,
+            validationErrors: errors.array()
+        });
+    }
 
     Product.findById(prodId)
         .then(product => {
@@ -62,22 +114,38 @@ exports.postEditProduct = (req, res, next) => {
             product.title = updatedTitle;
             product.price = updatedPrice;
             product.description = updatedDesc;
-            product.imageUrl = updatedImageUrl;
+            if (image) {
+                fileHelper.deleteFile(product.imageUrl);
+                product.imageUrl = image.path;
+            }
             return product.save()
                 .then(() => {
                     res.redirect('/admin/products');
                 });
         })
-        .catch();
+        .catch((err) => {
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
 };
 
-exports.postDeleteProduct = (req, res, next) => {
-    const prodId = req.body.productId;
-    Product.deleteOne({ _id: prodId, userId: req.user._id })
-        .then(() => {
-            res.redirect('/admin/products');
+exports.deleteProduct = (req, res, next) => {
+    const prodId = req.params.productId;
+    Product.findById(prodId)
+        .then(product => {
+            if (!product) {
+                return next(new Error('Product not found.'));
+            }
+            fileHelper.deleteFile(product.imageUrl);
+            return Product.deleteOne({ _id: prodId, userId: req.user._id });
         })
-        .catch();
+        .then(() => {
+            res.status(200).json({ message: 'Success' });
+        })
+        .catch((err) => {
+            res.status(500).json({ message: 'Deleting product failed.' });
+        });
 };
 
 exports.getProducts = (req, res, next) => {
@@ -91,5 +159,9 @@ exports.getProducts = (req, res, next) => {
                 path: '/admin/products',
             });
         })
-        .catch();
+        .catch((err) => {
+            const error = new Error(err);
+            error.httpStatusCode = 500;
+            return next(error);
+        });
 };
